+++ platform/lego_nxt/cmds/bt_control/bt_control.c
@@ -8,9 +8,9 @@
 
 #include <stddef.h>
 #include <embox/cmd.h>
-#include <pnet/repo.h>
-#include <pnet/types.h>
-#include <pnet/graph.h>
+#include <pnet/core/repo.h>
+#include <pnet/core/types.h>
+#include <pnet/core/graph.h>
 #include <pnet/node/direct_comm.h>
 #include <drivers/bluetooth/blue_core4.h>
