+++ src/drivers/bluetooth/lego/nxt_bluetooth_usart.c
@@ -22,9 +22,9 @@
 #include <embox/unit.h>
 
 
-#include <pnet/core.h>
-#include <pnet/repo.h>
-#include <pnet/pnet_pack.h>
+#include <pnet/core/core.h>
+#include <pnet/core/repo.h>
+#include <pnet/pack/pnet_pack.h>
 #include <pnet/pack/pack_alone.h>
 
 #define RX_PIN  ((uint32_t) (1 << OPTION_GET(NUMBER,rx_pin)))
