+++ src/pnet/core/Mybuild
@@ -2,7 +2,19 @@ package embox.pnet
 
 module core {
 	@IncludeExport(path="pnet/core")
+	source "prior_path.h"
+	@IncludeExport(path="pnet/core")
+	source "proto.h"
+	@IncludeExport(path="pnet/core")
 	source "node.h"
+	@IncludeExport(path="pnet/core")
+	source "core.h"
+	@IncludeExport(path="pnet/core")
+	source "types.h"
+	@IncludeExport(path="pnet/core")
+	source "repo.h"
+	@IncludeExport(path="pnet/core")
+	source "graph.h"
 	source "node.c", "repo.c", "graph.c"
 
 	option number pnet_nodes_quantity
