+++ src/pnet/core/dev.c
@@ -10,11 +10,11 @@
 #include <string.h>
 #include <embox/unit.h>
 
-#include <pnet/types.h>
-#include <pnet/core.h>
+#include <pnet/core/types.h>
+#include <pnet/core/core.h>
 #include <pnet/core/node.h>
-#include <pnet/repo.h>
-#include <pnet/pnet_pack.h>
+#include <pnet/core/repo.h>
+#include <pnet/pack/pnet_pack.h>
 #include <pnet/core/dev.h>
 
 #include <net/netdevice.h>
