+++ src/pnet/core/repo.c
@@ -10,8 +10,8 @@
 #include <util/array.h>
 #include <string.h>
 
-#include <pnet/core.h>
-#include <pnet/repo.h>
+#include <pnet/core/core.h>
+#include <pnet/core/repo.h>
 #include <pnet/core/node.h>
 
 ARRAY_SPREAD_DEF(const struct pnet_module, __pnet_mod_repo);
