+++ src/pnet/net_entry.c
@@ -12,11 +12,11 @@
 #include <linux/list.h>
 #include <stdio.h>
 
-#include <pnet/prior_path.h>
-#include <pnet/core.h>
+#include <pnet/core/prior_path.h>
+#include <pnet/core/core.h>
 #include <pnet/core/node.h>
-#include <pnet/repo.h>
-#include <pnet/pnet_pack.h>
+#include <pnet/core/repo.h>
+#include <pnet/pack/pnet_pack.h>
 
 #include <kernel/sched/schedee_priority.h>
 #include <kernel/lthread/lthread.h>
