+++ src/pnet/node/dev_linux.c
@@ -13,11 +13,11 @@
 #include <embox/unit.h>
 #include <string.h>
 
-#include <pnet/core.h>
-#include <pnet/node.h>
-#include <pnet/repo.h>
-#include <pnet/pnet_pack.h>
-#include <pnet/graph.h>
+#include <pnet/core/core.h>
+#include <pnet/core/node.h>
+#include <pnet/core/repo.h>
+#include <pnet/pack/pnet_pack.h>
+#include <pnet/core/graph.h>
 
 #include <net/skbuff.h>
 #include <net/netdevice.h>
