+++ src/pnet/node/info_printer.c
@@ -6,10 +6,10 @@
  * @author Anton Kozlov
  */
 #include <stdio.h>
-#include <pnet/core.h>
-#include <pnet/node.h>
-#include <pnet/repo.h>
-#include <pnet/pnet_pack.h>
+#include <pnet/core/core.h>
+#include <pnet/core/node.h>
+#include <pnet/core/repo.h>
+#include <pnet/pack/pnet_pack.h>
 #include <net/skbuff.h>
 
 static int stamp = 1;
