+++ src/pnet/node/skb_printer.c
@@ -8,10 +8,10 @@
  */
 
 #include <stdio.h>
-#include <pnet/core.h>
-#include <pnet/node.h>
-#include <pnet/repo.h>
-#include <pnet/pnet_pack.h>
+#include <pnet/core/core.h>
+#include <pnet/core/node.h>
+#include <pnet/core/repo.h>
+#include <pnet/pack/pnet_pack.h>
 
 #include <net/skbuff.h>
 #include <net/netdevice.h>
