+++ src/pnet/pack/pack_standalone/Mybuild
@@ -1,6 +1,8 @@
 package embox.pnet.pack
 
 module PnetPackSimple {
+	@IncludeExport(path="pnet/pack")
+	source "pack_alone.h"
 	source "pack_alone.c"
 
 	option number pnet_pack_quantity
