+++ src/pnet/pack/pnet_pack_factory.c
@@ -10,8 +10,8 @@
 #include <util/array.h>
 #include <mem/objalloc.h>
 
-#include <pnet/pnet_pack.h>
-#include <pnet/types.h>
+#include <pnet/pack/pnet_pack.h>
+#include <pnet/core/types.h>
 
 ARRAY_SPREAD_DEF(const struct pnet_pack_desc, __pnet_pack_registry);
