+++ src/pnet/pnet_entry.c
@@ -9,8 +9,8 @@
 #include <errno.h>
 #include <assert.h>
 
-#include <pnet/prior_path.h>
-#include <pnet/core.h>
+#include <pnet/core/prior_path.h>
+#include <pnet/core/core.h>
 #include <pnet/core/node.h>
 
 int pnet_entry(struct pnet_pack *pack) {
